requirements.txt
beautifulsoup4==4.9.0
django-environ==0.4.5
psycopg2==2.8.5
selenium==3.141.0
soupsieve==2.0
urllib3==1.25.9
django-environ==0.4.5